lease_payment/requirements.txt
Django >= 3.0, < 4.0
djangorestframework >= 3.9.1
django-filter == 2.0.0
mysqlclient >= 2.0.3
django-mysql >= 3.9
django-cors-headers >= 3.5.0
pytest-django >= 4.4.0
